Split GuildTemplate types
