+++ packages/types/src/index.ts
@@ -8,6 +8,7 @@ export * from './discord/entitlements.js'
 export * from './discord/gateway.js'
 export * from './discord/guilds.js'
 export * from './discord/guildScheduledEvents.js'
+export * from './discord/guildTemplates.js'
 export * from './discord/interactions.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
